SdpaFreeFormat can remove free variables.
