+++ solvers/csdp_solver.cc
@@ -161,8 +161,8 @@ void SolveProgramThroughNullspaceApproach(
   std::vector<Eigen::SparseMatrix<double>> A_hat;
   Eigen::VectorXd rhs_hat, y_hat;
   Eigen::SparseQR<Eigen::SparseMatrix<double>, Eigen::COLAMDOrdering<int>> QR_B;
-  RemoveFreeVariableByNullspaceApproach(sdpa_free_format, &C_hat, &A_hat,
-                                        &rhs_hat, &y_hat, &QR_B);
+  sdpa_free_format.RemoveFreeVariableByNullspaceApproach(
+      &C_hat, &A_hat, &rhs_hat, &y_hat, &QR_B);
 
   // Now try to call CSDP to solve this problem.
   csdp::blockmatrix C_csdp;
@@ -239,52 +239,13 @@ void SolveProgramThroughTwoSlackVariablesApproach(
       "0, and constraint y_plus - y_minus = free_variable. This can "
       "introduce numerical problems to the solver. Consider providing a lower "
       "and/or upper bound for each decision variable.");
-  std::vector<internal::BlockInX> X_hat_blocks = sdpa_free_format.X_blocks();
-  X_hat_blocks.emplace_back(internal::BlockType::kDiagonal,
-                            2 * sdpa_free_format.num_free_variables());
+  std::vector<internal::BlockInX> X_hat_blocks;
+  std::vector<Eigen::SparseMatrix<double>> A_hat;
+  Eigen::SparseMatrix<double> C_hat;
+  sdpa_free_format.RemoveFreeVariableByTwoSlackVariablesApproach(
+      &X_hat_blocks, &A_hat, &C_hat);
   const int num_X_hat_rows =
       sdpa_free_format.num_X_rows() + 2 * sdpa_free_format.num_free_variables();
-  std::vector<std::vector<Eigen::Triplet<double>>> A_hat_triplets =
-      sdpa_free_format.A_triplets();
-  for (int j = 0; j < sdpa_free_format.num_free_variables(); ++j) {
-    for (Eigen::SparseMatrix<double>::InnerIterator it(sdpa_free_format.B(), j);
-         it; ++it) {
-      const int i = it.row();
-      // Add the entry in Âᵢ that multiplies with sⱼ.
-      A_hat_triplets[i].emplace_back(sdpa_free_format.num_X_rows() + j,
-                                     sdpa_free_format.num_X_rows() + j,
-                                     it.value());
-      A_hat_triplets[i].emplace_back(
-          sdpa_free_format.num_X_rows() +
-              sdpa_free_format.num_free_variables() + j,
-          sdpa_free_format.num_X_rows() +
-              sdpa_free_format.num_free_variables() + j,
-          -it.value());
-    }
-  }
-  std::vector<Eigen::SparseMatrix<double>> A_hat;
-  A_hat.reserve(sdpa_free_format.A().size());
-  for (int i = 0; i < static_cast<int>(sdpa_free_format.A().size()); ++i) {
-    A_hat.emplace_back(num_X_hat_rows, num_X_hat_rows);
-    A_hat.back().setFromTriplets(A_hat_triplets[i].begin(),
-                                 A_hat_triplets[i].end());
-  }
-  // Add the entry in Ĉ that multiplies with sᵢ
-  std::vector<Eigen::Triplet<double>> C_hat_triplets =
-      sdpa_free_format.C_triplets();
-  for (Eigen::SparseMatrix<double>::InnerIterator it(sdpa_free_format.d(), 0);
-       it; ++it) {
-    const int i = it.row();
-    C_hat_triplets.emplace_back(sdpa_free_format.num_X_rows() + i,
-                                sdpa_free_format.num_X_rows() + i, it.value());
-    C_hat_triplets.emplace_back(sdpa_free_format.num_X_rows() +
-                                    sdpa_free_format.num_free_variables() + i,
-                                sdpa_free_format.num_X_rows() +
-                                    sdpa_free_format.num_free_variables() + i,
-                                -it.value());
-  }
-  Eigen::SparseMatrix<double> C_hat(num_X_hat_rows, num_X_hat_rows);
-  C_hat.setFromTriplets(C_hat_triplets.begin(), C_hat_triplets.end());
 
   // Now try to call CSDP to solve this problem.
   csdp::blockmatrix C_csdp;
@@ -359,14 +320,15 @@ void CsdpSolver::DoSolve(const MathematicalProgram& prog,
     SolveProgramWithNoFreeVariables(prog, sdpa_free_format, result);
   } else {
     switch (method_) {
-      case RemoveFreeVariableMethod::kNullspace: {
+      case internal::SdpaFreeFormat::RemoveFreeVariableMethod::kNullspace: {
         SolveProgramThroughNullspaceApproach(prog, sdpa_free_format,
                                                        result);
         break;
       }
-      case RemoveFreeVariableMethod::kTwoSlackVariables: {
-         SolveProgramThroughTwoSlackVariablesApproach(
-            prog, sdpa_free_format, result);
+      case internal::SdpaFreeFormat::RemoveFreeVariableMethod::
+          kTwoSlackVariables: {
+        SolveProgramThroughTwoSlackVariablesApproach(prog, sdpa_free_format,
+                                                     result);
         break;
       }
     }
