+++ solvers/sdpa_free_format.h
@@ -115,6 +115,59 @@ class SdpaFreeFormat {
   explicit SdpaFreeFormat(const MathematicalProgram& prog);
 
   ~SdpaFreeFormat();
+  /**
+   * SDPA format doesn't accept free variables, namely the problem it solves is
+   * in this form P1
+   *
+   *     max tr(C * X)
+   *     s.t tr(Aᵢ*X) = aᵢ
+   *         X ≽ 0.
+   *
+   * Notice that the decision variable X has to be in the proper cone X ≽ 0, and
+   * it doesn't accept free variable (without the conic constraint). On the
+   * other hand, most real-world applications require free variables, namely
+   * problems in this form P2
+   *
+   *     max tr(C * X) + dᵀs
+   *     s.t tr(Aᵢ*X) + bᵢᵀs = aᵢ
+   *         X ≽ 0
+   *         s is free.
+   *
+   * In order to remove the free variables, we consider two approaches.
+   * 1. Replace a free variable s with two variables s = p - q, p ≥ 0, q ≥ 0.
+   * 2. First write the dual of the problem P2 as D2
+   *
+   *        min aᵀy
+   *        s.t ∑ᵢ yᵢAᵢ - C = Z
+   *            Z ≽ 0
+   *            Bᵀ * y = d,
+   *
+   *    where bᵢᵀ is the i'th row of B.
+   *    The last constraint Bᵀ * y = d means y = ŷ + Nt, where Bᵀ * ŷ = d, and N
+   *    is the null space of Bᵀ. Hence, D2 is equivalent to the following
+   *    problem, D3
+   *
+   *        min aᵀNt + aᵀŷ
+   *        s.t ∑ᵢ tᵢFᵢ - (C -∑ᵢ ŷᵢAᵢ) = Z
+   *            Z ≽ 0,
+   *
+   *    where Fᵢ  = ∑ⱼ NⱼᵢAⱼ. D3 is the dual of the following primal problem P3
+   *    without free variables
+   *
+   *        max tr((C-∑ᵢ ŷᵢAᵢ)*X̂) + aᵀŷ
+   *        s.t tr(FᵢX̂) = (Nᵀa)(i)
+   *            X̂ ≽ 0.
+   *
+   *    Then (X, s) = (X̂, B⁻¹(a - tr(Aᵢ X̂))) is the solution to the original
+   *    problem P2.
+   *
+   */
+  enum RemoveFreeVariableMethod {
+    kTwoSlackVariables,  ///< Approach 1, replace a free variable s as
+                         ///< s = y⁺ - y⁻, y⁺ ≥ 0, y⁻ ≥ 0.
+    kNullspace,  ///< Approach 2, reformulate the dual problem by considering
+                 ///< the nullspace of the linear constraint in the dual.
+  };
 
   const std::vector<BlockInX>& X_blocks() const { return X_blocks_; }
 
@@ -161,6 +214,60 @@ class SdpaFreeFormat {
    * MathematicalProgram does. We store the constant cost term here.
    */
   double constant_min_cost_term() const { return constant_min_cost_term_; }
+
+  /**
+   * For the following problem
+   *
+   *     max tr(C * X) + dᵀs
+   *     s.t tr(Aᵢ*X) + bᵢᵀs = aᵢ
+   *         X ≽ 0
+   *         s is free,
+   *
+   * remove the free variable s by considering the nullspace of Bᵀ, where bᵢᵀ
+   * is the i'th row of B.
+   *
+   *     max tr((C-∑ᵢ ŷᵢAᵢ)*X̂) + aᵀŷ
+   *     s.t tr(FᵢX̂) = (Nᵀa)(i)
+   *         X̂ ≽ 0,
+   * where Fᵢ  = ∑ⱼ NⱼᵢAⱼ, N is the null space of Bᵀ. Bᵀ * ŷ = d.
+   * For more information, refer to CsdpSolver::RemoveFreeVariableMethod for the
+   * derivation.
+   * @param C_hat[out] C_hat is (C-∑ᵢ ŷᵢAᵢ) in the documentation.
+   * @param A_hat[out] A_hat[i] is Fᵢ in the documentation.
+   * @param rhs_hat[out] rhs_hat is (Nᵀa)
+   * @param y_hat[out] ŷ in the documentation.
+   * @param QR_B[out] The QR decomposition of B.
+   */
+  void RemoveFreeVariableByNullspaceApproach(
+      Eigen::SparseMatrix<double>* C_hat,
+      std::vector<Eigen::SparseMatrix<double>>* A_hat, Eigen::VectorXd* rhs_hat,
+      Eigen::VectorXd* y_hat,
+      Eigen::SparseQR<Eigen::SparseMatrix<double>, Eigen::COLAMDOrdering<int>>*
+          QR_B) const;
+
+  /**
+   * For the problem
+   * max tr(C * X) + dᵀs
+   * s.t tr(Aᵢ*X) + bᵢᵀs = aᵢ
+   *     X ≽ 0
+   *     s is free.
+   * Remove the free variable s by introducing two slack variables y⁺ ≥ 0 and y⁻
+   * ≥ 0, and the constraint y⁺ - y⁻ = s. We get a new program without free
+   * variables.
+   * max tr(Ĉ * X̂)
+   * s.t tr(Âᵢ*X̂) = aᵢ
+   *     X̂ ≽ 0
+   * where Ĉ = diag(C, diag(d), -diag(d))
+   *       X̂ = diag(X, diag(y⁺), diag(y⁻))
+   *       Âᵢ = diag(Aᵢ, diag(bᵢ), -diag(bᵢ))
+   * @param[out] X_hat_blocks The block matrix recording X̂.
+   * @param[out] A_hat A_hat[i] is Âᵢ in the documentation.
+   * @param[out] C_hat Ĉ in the documentation.
+   */
+  void RemoveFreeVariableByTwoSlackVariablesApproach(
+      std::vector<internal::BlockInX>* X_hat_blocks,
+      std::vector<Eigen::SparseMatrix<double>>* A_hat,
+      Eigen::SparseMatrix<double>* C_hat) const;
 
  private:
   // Go through all the positive semidefinite constraint in @p prog, and
